requirements.txt
azure-identity
azure-eventhub
requests
python-dotenv